Add c64.asm built-in macro library + download PRG button
The macro library can be included in the main assembly files as
"c64.asm".  The main motivation for this stdlib is to support the
BASIC startup sequence for working .prg file output.

Updates c64jasm bundle to latest in GH c64jasm HEAD.

Add "Download .prg" button that can be used to download the
resulting PRG.